+++ Magic/CHANGELOG.md
@@ -16,6 +16,7 @@
 ## 6.9.4
 
  - Fix compatibility with PerWorldInventory plugin via a "deactivate_wand_on_world_change: true" config.yml option
+ - Put the Heroes wand back, I removed it prematurely.
 
 ## 6.9.3